I made a file to run on the Arduino and the Raspberry Pi
The file names are stupid. Please change them. In the arduino code, the happyStuff() function runs when the bot sees a happy face. The sadStuff() function runs whenever the bot sees a sad face. I have no way to test whether or not the above statements are true, so test those.